+++ Wikimate.php
@@ -526,6 +526,7 @@ class WikiPage
 	public function getText($refresh = false)
 	{
 		if ($refresh) { // We want to query the API
+			// Specify relevant page properties to retrieve
 			$data = array(
 				'titles' => $this->title,
 				'prop' => 'info|revisions',
@@ -1053,7 +1054,7 @@ class WikiFile
 	public function getInfo($refresh = false)
 	{
 		if ($refresh) { // We want to query the API
-
+			// Specify relevant file properties to retrieve
 			$data = array(
 				'titles' => 'File:' . $this->filename,
 				'prop' => 'info|imageinfo',
@@ -1115,7 +1116,7 @@ class WikiFile
 	/**
 	 * Returns the bit depth of this file.
 	 *
-	 * @return  int  The bit depth of this file
+	 * @return  integer  The bit depth of this file
 	 */
 	public function getBitDepth()
 	{
@@ -1175,7 +1176,7 @@ class WikiFile
 	/**
 	 * Returns the height of this file.
 	 *
-	 * @return  int  The height of this file
+	 * @return  integer  The height of this file
 	 */
 	public function getHeight()
 	{
@@ -1235,7 +1236,7 @@ class WikiFile
 	/**
 	 * Returns the size of this file.
 	 *
-	 * @return  int  The size of this file
+	 * @return  integer  The size of this file
 	 */
 	public function getSize()
 	{
@@ -1286,7 +1287,7 @@ class WikiFile
 	/**
 	 * Returns the ID of the user who uploaded this file.
 	 *
-	 * @return  int  The user ID of this file
+	 * @return  integer  The user ID of this file
 	 */
 	public function getUserId()
 	{
@@ -1296,7 +1297,7 @@ class WikiFile
 	/**
 	 * Returns the width of this file.
 	 *
-	 * @return  int  The width of this file
+	 * @return  integer  The width of this file
 	 */
 	public function getWidth()
 	{
@@ -1315,7 +1316,7 @@ class WikiFile
 	 *
 	 * @return  mixed  Contents (string), or null if error
 	 */
-	public function download()
+	public function downloadData()
 	{
 		// Download file, or handle error
 		$data = $this->wikimate->download($this->getUrl());
@@ -1337,7 +1338,7 @@ class WikiFile
 	public function downloadFile($path)
 	{
 		// Download contents of current file
-		if (($data = $this->download()) === null) {
+		if (($data = $this->downloadData()) === null) {
 			return false;
 		}
 
@@ -1352,18 +1353,18 @@ class WikiFile
 	}
 
 	/**
-	 * Uploads the given contents to the current file.
-	 * $text is only used for the article page of a new file, not an existing
-	 * (update that via WikiPage::setText()).
+	 * Uploads to the current file using the given parameters.
+	 * $text is only used for the page contents of a new file,
+	 * not an existing (update that via WikiPage::setText()).
 	 * If no $text is specified, $comment will be used as new page text.
 	 *
-	 * @param   string   $data       The data to upload
+	 * @param   array    $params     The upload parameters
 	 * @param   string   $comment    Upload comment for the file
 	 * @param   string   $text       The article text for the file page
 	 * @param   boolean  $overwrite  True to overwrite existing file
 	 * @return  boolean              True if uploading was successful
 	 */
-	public function upload($data, $comment, $text = null, $overwrite = false)
+	private function uploadCommon(array $params, $comment, $text = null, $overwrite = false)
 	{
 		// Check whether to overwrite existing file
 		if ($this->exists && !$overwrite) {
@@ -1373,15 +1374,12 @@ class WikiFile
 		}
 
 		// Collect upload parameters
-		$params = array(
-			'filename' => $this->filename,
-			'comment' => $comment,
-			'ignorewarnings' => $overwrite,
-			'file' => $data,
-			'token' => $this->edittoken,
-		);
-		if ($text !== null) {
-			$params['text'] = $text;
+		$params['filename']       = $this->filename;
+		$params['comment']        = $comment;
+		$params['ignorewarnings'] = $overwrite;
+		$params['token']          = $this->edittoken;
+		if (!is_null($text)) {
+			$params['text']   = $text;
 		}
 
 		// Upload file, or handle error
@@ -1406,9 +1404,32 @@ class WikiFile
 	}
 
 	/**
+	 * Uploads the given contents to the current file.
+	 * $text is only used for the page contents of a new file,
+	 * not an existing (update that via WikiPage::setText()).
+	 * If no $text is specified, $comment will be used as new page text.
+	 *
+	 * @param   string   $data       The data to upload
+	 * @param   string   $comment    Upload comment for the file
+	 * @param   string   $text       The article text for the file page
+	 * @param   boolean  $overwrite  True to overwrite existing file
+	 * @return  boolean              True if uploading was successful
+	 */
+	public function uploadData($data, $comment, $text = null, $overwrite = false)
+	{
+		// Collect upload parameter
+		$params = array(
+			'file' => $data,
+		);
+
+		// Upload contents to current file
+		return $this->uploadCommon($params, $comment, $text, $overwrite);
+	}
+
+	/**
 	 * Reads contents from the given path and uploads it to the current file.
-	 * $text is only used for the article page of a new file, not an existing
-	 * (update that via WikiPage::setText()).
+	 * $text is only used for the page contents of a new file,
+	 * not an existing (update that via WikiPage::setText()).
 	 * If no $text is specified, $comment will be used as new page text.
 	 *
 	 * @param   string   $path       The file path to upload
@@ -1427,6 +1448,29 @@ class WikiFile
 		}
 
 		// Upload contents to current file
-		return $this->upload($data, $comment, $text, $overwrite);
+		return $this->uploadData($data, $comment, $text, $overwrite);
+	}
+
+	/**
+	 * Uploads file contents from the given URL to the current file.
+	 * $text is only used for the page contents of a new file,
+	 * not an existing (update that via WikiPage::setText()).
+	 * If no $text is specified, $comment will be used as new page text.
+	 *
+	 * @param   string   $url        The URL from which to upload
+	 * @param   string   $comment    Upload comment for the file
+	 * @param   string   $text       The article text for the file page
+	 * @param   boolean  $overwrite  True to overwrite existing file
+	 * @return  boolean              True if uploading was successful
+	 */
+	public function uploadFromUrl($url, $comment, $text = null, $overwrite = false)
+	{
+		// Collect upload parameter
+		$params = array(
+			'url' => $url,
+		);
+
+		// Upload URL to current file
+		return $this->uploadCommon($params, $comment, $text, $overwrite);
 	}
 }